send email fix

+++ frontEnd/src/common/navbar.ts
@@ -104,7 +104,9 @@ export async function validateUser() {
     } catch (error) {
       if (error instanceof AxiosError && error.response) {
         if (error.response.status == 401) {
-          localStorage.clear();
+          if (accessToken) {
+            localStorage.clear();
+          }
           window.location.href = "/";
         }
       }
